refactor: `revoke-permission-button` to use `useSendTransaction` hook
